Improve over the last version by depending on the latest snapshot.
The latest SNAPSHOT on main includes a `Store` facility that
does not track direct dependencies between components; it does
not recursively instantiate required modules, it just assumes
that instantiating a given module at a given moment means
that all the requirements are satisfied at that time.

The missing bit to turning the Store into a linker
is a layer that tracks dependencies, linearizes them into
a sequence, and then instantiate them in order.
We also track cycles, introducing an indirection (a trampoline)
to break out of them.

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>